Pi Zero performance: thread SDS011, persistent log, float, -O2, config fix
Closes #17, closes #18, closes #19, closes #20, closes #21, closes #22

SDS011 threaded reader (#17):
- SDS011Reader class (same RAII pattern as VEDirectReader): one thread
  per device reads frames byte-by-byte with checksum validation,
  updates snapshot under mutex; main loop calls getReadings() without
  blocking; eliminates 2s stall per poll cycle on single-core Pi Zero
- On reload, readers are destroyed and recreated like VEDirect readers

Persistent log file (#18, #19):
- g_logStream opened once at startup via openLogFile(), protected by
  g_logMutex for thread-safe writes from sensor reader threads
- logMessage() writes to the open stream directly — no open/close per
  call, no ensureDirectoriesExist() on every write
- openLogFile() called again on SIGHUP if log path changes

CMake optimisation (#20):
- -O2 added for all non-debug builds
- RPI_ZERO option adds -march=armv6 -mfpu=vfp -mfloat-abi=hard

float for SensorReading::value (#21):
- SensorReading::value, SDS011Reading fields, readDSTemperature(),
  readEE895(), readINA219() and all derived calculations changed to
  float; VEDirect parseFrame uses stof; eliminates software double
  emulation on ARMv6 VFPv2

ConfigParser singleton fixed (#22):
- loadServiceConfig() now uses a local ConfigParser instance instead
  of the global singleton, so each reload starts from a clean slate
  and removed config keys do not persist

+++ include/config.hpp
@@ -69,7 +69,7 @@ namespace rpi {
         int ina219I2CBus;
         int ina219I2CAddress;
         std::string ina219SensorId;
-        double ina219ShuntResistance;
+        float ina219ShuntResistance;
         bool vedirectEnabled;
         std::vector<std::string> vedirectDevices;
         std::vector<std::string> vedirectSensorIds;
